requirements.txt
allure-pytest==2.14.1
allure-python-commons==2.14.1
attrs==23.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.5
pytest-trio==0.8.0
pytest-rerunfailures==15.0
python-dotenv==1.0.0
requests==2.31.0
selenium==4.31.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.10.3
urllib3==2.0.4
webdriver-manager==4.0.0
wsproto==1.2.0
